FROMGIT: perf/x86/intel: Set PERF_ATTACH_SCHED_CB for large PEBS and LBR
To supply a PID/TID for large PEBS, it requires flushing the PEBS buffer
in a context switch.

For normal LBRs, a context switch can flip the address space and LBR
entries are not tagged with an identifier, we need to wipe the LBR, even
for per-cpu events.

For LBR callstack, save/restore the stack is required during a context
switch.

Set PERF_ATTACH_SCHED_CB for the event with large PEBS & LBR.

BUG=b:170558957
TEST=build/boot

Fixes: 9c964efa4330 ("perf/x86/intel: Drain the PEBS buffer during context switches")
Signed-off-by: Kan Liang <kan.liang@linux.intel.com>
Signed-off-by: Peter Zijlstra (Intel) <peterz@infradead.org>
Signed-off-by: Ingo Molnar <mingo@kernel.org>
Link: https://lkml.kernel.org/r/20201130193842.10569-2-kan.liang@linux.intel.com
Signed-off-by: Sonny Rao <sonnyrao@chromium.org>
(cherry picked from commit afbef30149587ad46f4780b1e0cc5e219745ce90
 git://git.kernel.org/pub/scm/linux/kernel/git/tip/tip.git perf/urgent)
Change-Id: I6febe937694ac2d337b257cfdcc143e6050b644a
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2746749
Reviewed-by: Sean Paul <seanpaul@chromium.org>