CHROMIUM: nvme: disable the write zeros command for Pixelbook.
This is a followup to a suggestion made in https://crbug.com/1141713#c68

This is a chromium-only patch because we aim to test the theory that
the "Write Zeros" command is causing a problem on Pixelbooks in our lab.
The evidence still seems a little too circumstantial at the moment
(to me) to proactively do the leg work to get it upstream given
there's a risk that we will want to revert if the issue remains, but if
it turns out to mitigate the issue then for sure I will revert and
get it in upstream instead. If you think it's better off going upstream
in the first place please let me know.

I retrieved the device info from:
http://cs/chromeos_internal/src/platform/chromeos-hwid/v3/EVE

BUG=b:174236852
TEST=just compiled (emerge-eve-kernelnext chromeos-kernel-5_4)

Change-Id: Iefe61257340139c157e34bfec2e885a19ee7a0e8
Signed-off-by: Kevin Chowski <chowski@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2566697
Tested-by: Kevin Chowski <chowski@google.com>
Reviewed-by: Gwendal Grignou <gwendal@chromium.org>
Commit-Queue: Kevin Chowski <chowski@google.com>
Auto-Submit: Kevin Chowski <chowski@google.com>