FROMLIST: iommu/mediatek: Add a flag for iova 34bits case
Add a HW flag for if the HW support 34bit IOVA. the previous SoC
still use 32bit. normally the lvl1 pgtable size is 16KB when ias == 32.
if ias == 34, lvl1 pgtable size is 16KB * 4. The purpose of this patch
is to save 16KB*3 continuous memory for the previous SoC.

Signed-off-by: Yong Wu <yong.wu@mediatek.com>
(am from https://lore.kernel.org/patchwork/patch/1362969/)

BUG=b:153618578
TEST=Boot OK

Signed-off-by: Yong Wu <yong.wu@mediatek.com>
Change-Id: I3573509c6b49dbb8b71d9f6aff2e30a40d35d645
Signed-off-by: Nicolas Boichat <drinkcat@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2624309
Reviewed-by: Hsin-Yi Wang <hsinyi@chromium.org>