FROMLIST: drm/i915/pxp: Introduce Intel PXP component
PXP (Protected Xe Path) is an i915 componment, available on GEN12+,
that helps to establish the hardware protected session and manage
the status of the alive software session, as well as its life cycle.

This patch series is to allow the kernel space to create and
manage a single hardware session (a.k.a default session or
arbitrary session). So Mesa can allocate the protected buffer,
which is encrypted with the leverage of the arbitrary hardware
session.

Signed-off-by: Huang, Sean Z <sean.z.huang@intel.com>
(am from https://patchwork.freedesktop.org/patch/407032/)
(also found at https://lore.kernel.org/r/20201211090457.32674-2-sean.z.huang@intel.com)

BUG=b:155511178
TEST=None

Signed-off-by: Sean Paul <seanpaul@chromium.org>
Change-Id: I6eed8a93dcd20754a2c48eda741ac8759886109b
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2587327
Reviewed-by: Stéphane Marchesin <marcheu@chromium.org>