UPSTREAM: media: venus: use contig vb2 ops
This driver uses the SG vb2 ops, but effectively only ever accesses the
first entry of the SG table, indicating that it expects a flat layout.
Switch it to use the contiguous ops to make sure this expected invariant
is always enforced. Since the device is supposed to be behind an IOMMU
this should have little to none practical consequences beyond making the
driver not rely on a particular behavior of the SG implementation.

Reported-by: Tomasz Figa <tfiga@chromium.org>
Signed-off-by: Alexandre Courbot <acourbot@chromium.org>
Signed-off-by: Stanimir Varbanov <stanimir.varbanov@linaro.org>
Signed-off-by: Mauro Carvalho Chehab <mchehab+huawei@kernel.org>
(cherry picked from commit cc82fd691a3a7e8b46bd94fe7cacb2835015df22)

BUG=b:149115731
TEST=Run Camera app in lazor
TEST=Try https://appr.tc/?debug=loopback&vsc=vp8

Signed-off-by: Ricardo Ribalda <ribalda@chromium.org>
Change-Id: I088deac81fb5cd249f81f3e246d37f6e5a0e0fc8
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2729002
Reviewed-by: Sean Paul <seanpaul@chromium.org>
Reviewed-by: Douglas Anderson <dianders@chromium.org>
Reviewed-by: Alexandre Courbot <acourbot@chromium.org>