FROMLIST: mac80211: Add support to trigger sta disconnect on hardware restart
Currently in case of target hardware restart, we just reconfig and
re-enable the security keys and enable the network queues to start
data traffic back from where it was interrupted.

Many ath10k wifi chipsets have sequence numbers for the data
packets assigned by firmware and the mac sequence number will
restart from zero after target hardware restart leading to mismatch
in the sequence number expected by the remote peer vs the sequence
number of the frame sent by the target firmware.

This mismatch in sequence number will cause out-of-order packets
on the remote peer and all the frames sent by the device are dropped
until we reach the sequence number which was sent before we restarted
the target hardware

In order to fix this, we trigger a sta disconnect, for the targets
which expose this corresponding wiphy flag, in case of target hw
restart. After this there will be a fresh connection and thereby
avoiding the dropping of frames by remote peer.

The right fix would be to pull the entire data path into the host
which is not feasible or would need lots of complex changes and
will still be inefficient.

Tested on ath10k using WCN3990, QCA6174

Signed-off-by: Youghandhar Chintala <youghand@codeaurora.org>
(am from https://patchwork.kernel.org/patch/11975371/)
(also found at https://lore.kernel.org/r/20201215172352.5311-1-youghand@codeaurora.org)

BUG=b:156060940
TEST= On soft FW crash, the device recovers consistently.

Signed-off-by: Abhishek Kumar <kuabhs@chromium.org>
Change-Id: I472f8ed6ac7e0214813c3fff9e2f8e2b3fe3b6d1
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2601942
Reviewed-by: Sean Paul <seanpaul@chromium.org>
Reviewed-by: Brian Norris <briannorris@chromium.org>