FROMGIT: spi: spi-geni-qcom: Print an error when we timeout setting the CS
If we're using geni to manage the chip select line (don't do it--use a
GPIO!) and we happen to get a timeout waiting for the chip select
command to be completed, no errors are printed even though things
might not be in the best shape.  Let's add a print.

Signed-off-by: Douglas Anderson <dianders@chromium.org>
Reviewed-by: Stephen Boyd <swboyd@chromium.org>
Link: https://lore.kernel.org/r/20201217142842.v3.4.I666b37646de9652cef438ac7c2c6c2053367fc6b@changeid
Signed-off-by: Mark Brown <broonie@kernel.org>
(cherry picked from commit 17fa81aa702ec118f2b835715897041675b06336
 git://git.kernel.org/pub/scm/linux/kernel/git/broonie/spi.git for-5.11)

BUG=b:175219045
TEST=Hack long irq latency and no crash

Change-Id: I666b37646de9652cef438ac7c2c6c2053367fc6b
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2611493
Tested-by: Douglas Anderson <dianders@chromium.org>
Reviewed-by: Sean Paul <seanpaul@chromium.org>
Reviewed-by: Stephen Boyd <swboyd@chromium.org>
Commit-Queue: Douglas Anderson <dianders@chromium.org>