FROMGIT: net: ipa: use state to determine event ring command success
This patch implements the same basic fix for event rings as the
previous one does for channels.

The result of issuing an event ring control command should be that
the event ring changes state.  If enabled, a completion interrupt
signals that the event ring state has changed.  This interrupt is
enabled by gsi_evt_ring_command() and disabled again after the
command has completed (or we time out).

There is a window of time during which the command could complete
successfully without interrupting.  This would cause the event ring
to transition to the desired new state.

So whether a event ring command ends via completion interrupt or
timeout, we can consider the command successful if the event ring
has entered the desired state (and a failure if it has not,
regardless of the cause).

Fixes: b4175f8731f78 ("net: ipa: only enable GSI event control IRQs when needed")
Signed-off-by: Alex Elder <elder@linaro.org>
Signed-off-by: Jakub Kicinski <kuba@kernel.org>
(cherry picked from commit 428b448ee764a264b7a2eeed295b282755114aa7
 https://git.kernel.org/pub/scm/linux/kernel/git/netdev/net.git master)

BUG=b:175365709
TEST=Modem functional, no IPA hang

Signed-off-by: Sujit Kautkar <sujitka@chromium.org>
Change-Id: If45abcb2f0cac713ccfad5faddcf614181eed9c4
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2602546
Reviewed-by: Matthias Kaehlcke <mka@chromium.org>
Reviewed-by: Sean Paul <seanpaul@chromium.org>