BACKPORT: FROMLIST: iommu/mediatek: Add get_domain_id from dev->dma_range_map
Add a new interface _get_domain_id from dev->dma_range_map,
The iommu consumer device will use dma-ranges in dtsi node to indicate
its dma address region requirement. In this iommu driver, we will get
the requirement and decide which iova domain it should locate.

In the lastest SoC, there will be several iova-regions(domains), we will
compare and calculate which domain is right. If the start/end of device
requirement equal some region. it is best fit of course. If it is inside
some region, it is also ok. the iova requirement of a device should not
be inside two or more regions.

Signed-off-by: Yong Wu <yong.wu@mediatek.com>
(am from https://lore.kernel.org/patchwork/patch/1362983/)
[drinkcat: chromeos-5.4 does not have dev->dma_range_map yet, stub
mtk_iommu_get_domain_id to return 0 for now, properly backporting
this will be tracked in b:178756074]

BUG=b:178756074
BUG=b:153618578
TEST=Boot OK

Signed-off-by: Yong Wu <yong.wu@mediatek.com>
Change-Id: Ided61e943f1ad02bf23ad6a731b8ad63625bede4
Signed-off-by: Nicolas Boichat <drinkcat@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2624314