FROMLIST: net: ipa: introduce atomic channel STOPPING flag
Introduce a new atomic flag bit to communicate that a channel is
stopping.  At the end of the NAPI poll loop, we normally re-enable
the IEOB interrupt, but now we won't do that if the channel is being
stopped.  This is required for the next patch.

Fixes: 650d1603825d8 ("soc: qcom: ipa: the generic software interface")
Signed-off-by: Alex Elder <elder@linaro.org>
(am from https://patchwork.kernel.org/patch/12004967/)
(also found at https://lore.kernel.org/r/20210107214325.7077-2-elder@linaro.org)

BUG=b:175365709
TEST=Booted trogdor, modem call functional, No IPA hang

Signed-off-by: Sujit Kautkar <sujitka@chromium.org>
Change-Id: Icbc0eacb9e050650cd94db991fe35ea7b28c26be
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2617180
Reviewed-by: Sean Paul <seanpaul@chromium.org>
Reviewed-by: Douglas Anderson <dianders@chromium.org>