UPSTREAM: thunderbolt: Add support for maxhopid XDomain property
USB4 inter-domain spec mandates that the compatible hosts expose a new
property "maxhopid" that tells the connection manager on the other side
what is the maximum supported input HopID over the connection. Since
this is depend on the lane adapter the cable is connected it needs to be
filled in dynamically.

For this reason we take a copy of the global properties and fill then
for each XDomain connection upon first connect, and then keep updating
it if the generation changes as services are being added/removed. We
also take advantage of this copy to fill in the hostname.

We also expose this maxhopid as an attribute under each XDomain device.

While there drop kernel-doc entry for property_lock which seems to be
left there when the structure was originally introduced.

Signed-off-by: Mika Westerberg <mika.westerberg@linux.intel.com>
(cherry picked from commit 46b494f286812a88caba28dd0810cf3a55747431)

BUG=b:182447303
TEST=Enable CONFIG_DYNAMIC_DEBUG=y & set thunderbolt.dyndbg, then
connect a peer device (i.e. Ubuntu laptop with TBT support), then
check dmesg for updated logs:
[   82.696959] thunderbolt 0-3: requesting remote UUID
[   82.697240] thunderbolt 0-3: got remote UUID
cf030000-0070-6d18-a3ec-744b59c2e21d
[   82.735263] thunderbolt 0000:00:0d.2: 3: received XDomain request 0x5
[   82.800898] thunderbolt 0-3: sending properties changed notification
[   83.635233] thunderbolt 0000:00:0d.2: 3: received XDomain request 0x3
[   83.744386] thunderbolt 0-3: requesting remote properties
[   83.746323] thunderbolt 0-3: new host found, vendor=0x8086 device=0x1
[   83.753530] thunderbolt 0-3: Intel Corp. cgbowman-MOBL3
[   83.770099] thunderbolt 0000:00:0d.2: allocating TX ring -1 of size 256
[   83.770124] thunderbolt 0000:00:0d.2: allocating RX ring -1 of size 256
[   83.818342] thunderbolt 0-3: sending properties changed notification

Signed-off-by: Casey Bowman <casey.g.bowman@intel.com>
Change-Id: Idd4fc38955137920512aec4675fc6a5e49145f29
Signed-off-by: Chintan Patel <chintan.m.patel@intel.corp-partner.google.com>
Signed-off-by: Daniel Campello <campello@chromium.org>