CHROMIUM: drm/virtio: add type and blob_mem to res info
type is an input.  When it is VIRTGPU_RESOURCE_INFO_TYPE_DEFAULT, the
ioctl returns the extended info (strides, etc.) until chromeos-5.4 and
blob_mem since chromeos-5.10.  When it is
VIRTGPU_RESOURCE_INFO_TYPE_EXTENDED, the ioctl always returns the
extended info.

At the moment, the userspace (sommelier, minigbm, and mesa) only works
with chromeos-5.4 because chromeos-5.10 returns blob_mem instead of the
extended info.  The userspace also always sets type to 0 (that is,
VIRTGPU_RESOURCE_INFO_TYPE_DEFAULT).  What this means is that this
change makes no difference to the current userspace on chromeos-5.4.

However, the userspace should be fixed to set type to
VIRTGPU_RESOURCE_INFO_TYPE_EXTENDED if that is the intention.  The
change still makes no difference on chromeos-5.4, but will make sure the
extended info is returned on chromeos-5.10.

BUG=b:177238493
TEST=glxgears in vm

Change-Id: I7b815ade22fda5399dca860b041ff3a4a18a902a
Signed-off-by: Chia-I Wu <olv@google.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2625398