CHROMIUM: drm/panel-simple: Add vrefresh for N116BCA-EA1
Like several similar commits:

Upstream has removed the vrefresh field in commit 0425662fdf05 ("drm:
Nuke mode->vrefresh").  That CL not only adds support to calculate the
vrefresh dynamically in some cases but it also removes all users which
makes it really hard to backport (because there are more users
upstream than downstream).  For now, let's add the vrefresh back in
for N116BCA-EA1.  If later someone backports the CL removing
vrefresh it will be a super obvious compile-time error.

NOTE: if we skip this patch things seem to work for the most part, but
I'm a little worried that there's some bit of code somewhere that was
relying on it.  Best to take it.

BUG=b:176918319, b:172695850
TEST=Stress going between VT1 and VT2 w/ CL:2530131

Signed-off-by: Douglas Anderson <dianders@chromium.org>
Change-Id: I115af73783dcd0ae91d9f20c5a8132b12e92bf0a
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2633896
Reviewed-by: Rob Clark <robdclark@chromium.org>