FROMGIT: net: ipa: remove a remoteproc dependency
The IPA driver currently requires a DT property to be defined whose
value is the phandle for the modem subsystem.  This was needed to
look up a remoteproc structure pointer used when registering for
notifications in the original IPA notification mechanism.

Remoteproc provides a more generic SSR notifier system, and the IPA
driver switched over to it last summer, but this remoteproc phandle
dependency was not removed at that time.

Get rid of the IPA remoteproc pointer and stop requiring the phandle
be specified.

This avoids a link error (rproc_put() not defined) for certain
configurations.

Reported-by: Randy Dunlap <rdunlap@infradead.org>
Signed-off-by: Alex Elder <elder@linaro.org>
Signed-off-by: Jakub Kicinski <kuba@kernel.org>
(cherry picked from commit 86fdf1fc60e95ef4721de1b5f802f6a44db848c1
 https://git.kernel.org/pub/scm/linux/kernel/git/netdev/net-next.git master)

BUG=b:177097776
TEST=Modem functional, No IPA hang in suspend stress test

Signed-off-by: Sujit Kautkar <sujitka@chromium.org>
Change-Id: Iebb11175e8af8eb12741f6c25a9771ff45c03a03
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2674390
Reviewed-by: Sean Paul <seanpaul@chromium.org>
Reviewed-by: Evan Green <evgreen@chromium.org>