UPSTREAM: clk: divider: fix initialization with parent_hw
If a driver registers a divider clock with a parent_hw instead of the
parent_name, the parent_hw is ignored and the clock does not have a
parent.

Fix this by initializing the parents the same way they are initialized
for clock gates.

Fixes: ff258817137a ("clk: divider: Add support for specifying parents via DT/pointers")
Signed-off-by: Michael Tretter <m.tretter@pengutronix.de>
Reviewed-by: Stephen Boyd <sboyd@kernel.org>
Acked-by: Michal Simek <michal.simek@xilinx.com>
Link: https://lore.kernel.org/r/20210121071659.1226489-3-m.tretter@pengutronix.de
Signed-off-by: Stephen Boyd <sboyd@kernel.org>
(cherry picked from commit 0225daea08141b1dff681502d5af70b71e8b11ec)

BUG=b:162602229
TEST=Hang DPU on Lazor; see devcoredump

Signed-off-by: Stephen Boyd <swboyd@chromium.org>
Change-Id: I60a02452b5432857c1917c5c24fc70db5576d151
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2938325
Reviewed-by: Sean Paul <seanpaul@chromium.org>
Reviewed-by: Rob Clark <robdclark@chromium.org>