FROMGIT: net: ipa: count actual work done in gsi_channel_poll()
There is an off-by-one problem in gsi_channel_poll().  The count of
transactions completed is incremented each time through the loop
*before* determining whether there is any more work to do.  As a
result, if we exit the loop early the counter its value is one more
than the number of transactions actually processed.

Instead, increment the count after processing, to ensure it reflects
the number of processed transactions.  The result is more naturally
described as a for loop rather than a while loop, so change that.

Signed-off-by: Alex Elder <elder@linaro.org>
Signed-off-by: Jakub Kicinski <kuba@kernel.org>
(cherry picked from commit c80c4a1ea47f354584c8055015561c4f1ece8f7a
 https://git.kernel.org/pub/scm/linux/kernel/git/netdev/net-next.git master)

BUG=b:177097776
TEST=Modem functional, No IPA hang in suspend stress test

Signed-off-by: Sujit Kautkar <sujitka@chromium.org>
Change-Id: If78dfead578dc3b4394735e863d1e719ce05b126
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2674393
Reviewed-by: Sean Paul <seanpaul@chromium.org>
Reviewed-by: Evan Green <evgreen@chromium.org>