CHROMIUM: config: Re-enable POWER_ALLOCATOR for arm64
In CL:2601803, the patches robot suggested that we fix the
dependencies for THERMAL_DEFAULT_GOV_POWER_ALLOCATOR.  It's good to
have the dependencies right, but because of the change from "select"
to "depends on" and the way defconfigs work, this had the unfortunate
side effect of totally disabling THERMAL_DEFAULT_GOV_POWER_ALLOCATOR.
Oops.

Let's re-enable it.  While we're at it, it suddenly becomes obvious
that the rockchip64 config didn't have ENERGY_MODEL turned on because
without it THERMAL_DEFAULT_GOV_POWER_ALLOCATOR doesn't stay in
'arm64/common.config' and gets bubbled down to all configs except
rockchip64.  Let's turn that on.

This is:
  echo "CONFIG_THERMAL_GOV_POWER_ALLOCATOR=y" >> \
    chromeos/config/arm64/common.config
  echo "CONFIG_ENERGY_MODEL=y" >> \
    chromeos/config/arm64/chromiumos-rockchip64.flavour.config
  ./chromeos/scripts/kernelconfig olddefconfig
Atop a normalized config.

BUG=b:178099773
TEST=Grep configs

Signed-off-by: Douglas Anderson <dianders@chromium.org>
Change-Id: I98fd5419a2168a25e0c23097d00b6dd0c56c77bc
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2642887
Reviewed-by: Guenter Roeck <groeck@chromium.org>
Reviewed-by: Matthias Kaehlcke <mka@chromium.org>
Commit-Queue: Nicolas Boichat <drinkcat@chromium.org>