FIXUP: FROMLIST: power: supply: PCHG: Peripheral device charger
The problem appears to be that the driver registers for all cros_ec
events (host events, keyboard events, ...), but then if it ever gets
told about an event that's not a host event then it says "that's
illegal, everyone stop processing it" by returning NOTIFY_BAD.

Any driver later in the chain (like the keyboard driver) won't see
events.  On the system in front of me this means that all keys on the
system don't work anymore.

Fix to just say "DONE" which means "don't care".

Needs to be squashed into an upstream post.

BUG=b:178230365
TEST=Keys work again.

Signed-off-by: Douglas Anderson <dianders@chromium.org>
Change-Id: I74a135000748a5362906d158eda417f1c826346f
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2645556
Reviewed-by: Stephen Boyd <swboyd@chromium.org>
Reviewed-by: Alexandru M Stan <amstan@chromium.org>