CHROMIUM: thunderbolt: Prevent redundant retimer scans
Once the retimer scan is done for a given USB4 port, make a note of it,
so it's not repeated for a given USB4 port.

BUG=b:175045931
TEST=Boot with no devices attached to Type-C port and check sysfs
entries for retimer at /sys/bus/thunderbolt/device/0-0\:x.1 can be used
for updating retimer firmware.

TOREVERT=Once the upstream version of the patch lands in
maintainers tree, this patch should be reverted.
TOREVERT=b:180449562

Change-Id: Ie985e04e83097f1c7a21818ce45cd4fd94ca9ebe
Signed-off-by: Rajmohan Mani <rajmohan.mani@intel.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2646616
Tested-by: Kranthi Kuntala <kranthi.kuntala@intel.corp-partner.google.com>
Reviewed-by: Prashant Malani <pmalani@chromium.org>