CHROMIUM: thunderbolt: Modify tb_retimer_scan()
When updating the retimer firmware, when no devices are connected,
the USB4 retimer needs to be scanned and prepared for IO, for an
exising retimer device. Add a parameter to tb_retimer_scan() to
help achieve this.

BUG=b:175045931
TEST=Boot with no devices attached to Type-C port and check sysfs
entries for retimer at /sys/bus/thunderbolt/device/0-0\:x.1 can be used
for updating retimer firmware.

TOREVERT=Once the upstream version of the patch lands in
maintainers tree, this patch should be reverted.
TOREVERT=b:180449562

Change-Id: I4bee97ee91e64e80a422e006b72833750ef59508
Signed-off-by: Rajmohan Mani <rajmohan.mani@intel.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2646617
Tested-by: Kranthi Kuntala <kranthi.kuntala@intel.corp-partner.google.com>
Reviewed-by: Prashant Malani <pmalani@chromium.org>