CHROMIUM: thunderbolt: Improve tb_retimer_remove_all()
Currently all retimers including onboard retimers are removed,
when a switch is removed. Improved this by checking for a match of
the current switch being removed, against switch associated with
the USB4 port, to which the retimer is connected.

BUG=b:175045931
TEST=Boot with no devices attached to Type-C port, connect a
TBT/USB4 dock, remove and check if the onboard retimer sysfs
entries are still present.

TOREVERT=Once the upstream version of the patch lands in
maintainers tree, this patch should be reverted.
TOREVERT=b:180449562

Change-Id: I8715677b68dd935fd1d078f899a457a23c0e0268
Signed-off-by: Rajmohan Mani <rajmohan.mani@intel.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2648733
Tested-by: Kranthi Kuntala <kranthi.kuntala@intel.corp-partner.google.com>
Reviewed-by: Prashant Malani <pmalani@chromium.org>