FROMGIT: net: ipa: don't pass tag value to ipa_cmd_ip_tag_status_add()
We only send a tagged packet from the AP->command TX endpoint when
we're clearing the hardware pipeline.  And when we receive the
tagged packet we don't care what the actual tag value is.

Stop passing a tag value to ipa_cmd_ip_tag_status_add(), and just
encode 0 as the tag sent.  Fix the function that encodes the tag so
it uses the proper byte ordering.

Signed-off-by: Alex Elder <elder@linaro.org>
Acked-by: Willem de Bruijn <willemb@google.com>
Signed-off-by: Jakub Kicinski <kuba@kernel.org>
(cherry picked from commit 792b75b14786bae709124615a58a474ded15aa7c
 https://git.kernel.org/pub/scm/linux/kernel/git/netdev/net-next.git master)

BUG=b:177097776
TEST=Modem functional, No IPA hang in suspend stress test

Signed-off-by: Sujit Kautkar <sujitka@chromium.org>
Change-Id: I035a85ae964a5b2555637f1f138fc86e04b35484
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2674462
Reviewed-by: Sean Paul <seanpaul@chromium.org>
Reviewed-by: Evan Green <evgreen@chromium.org>