CHROMIUM: thunderbolt: Enable only onboard retimers when no devices are connected
The flows to enable communication to retimers, when no devices
are connected, is limited to onboard retimers. Modify this
function to achieve the same.

BUG=b:175045931
TEST=Boot with no devices attached to Type-C port and update retimer
firmware on a Type-C port and ensure Type-C based devices work fine
on all Type-C ports.

TOREVERT=Once the upstream version of the patch lands in
maintainers tree, this patch should be reverted.
TOREVERT=b:180449562

Change-Id: Icfccd9c1adff7bd0ef2891d470b6a3703b65d7d9
Signed-off-by: Rajmohan Mani <rajmohan.mani@intel.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2657916
Tested-by: Kranthi Kuntala <kranthi.kuntala@intel.corp-partner.google.com>
Reviewed-by: Prashant Malani <pmalani@chromium.org>